@genty_dataset is now chainable.
Can specifiy multiple @genty_dataset decorators on a method.
Various use cases. Allows for more flexible organization of
test cases. And there's a plan for a "defered" style of
datasets, where the actual params for the test case aren't
created until the test is called, as opposed to when the test
module is imported.